+++ src/common.hpp
@@ -13,7 +13,7 @@
 // ---------------------------------------------------------------------
 #define LIME_MAJOR_VERSION			0
 #define LIME_MINOR_VERSION			12
-#define LIME_BUILD					20260131
+#define LIME_BUILD					20260217
 #define LIME_YEAR					2026
 
 // ---------------------------------------------------------------------
@@ -21,8 +21,8 @@
 // ---------------------------------------------------------------------
 #define	VDC_RAM						0x1000000	// 16mb
 #define	VDC_RAM_MASK				(VDC_RAM-1)
-#define VDC_XRES					320
-#define VDC_YRES					176
+#define VDC_XRES					320			// 40 tiles
+#define VDC_YRES					176			// 22 tiles
 #define VDC_SCANLINES				(176+32)	// 176 visible, 32 vblank
 
 #define VDC_TILESET_ADDRESS			0x1000