+++ src/host.hpp
@@ -136,7 +136,6 @@ private:
 	SDL_FRect		viewer_texture_placement;
 
     inline uint32_t video_blend(uint32_t c0, uint32_t c1) {
-		// const uint8_t video_scanline_alpha = 0xb0;
 		return
 			((((c0 & 0x00ff00ff) + (c1 & 0x00ff00ff)) >> 1) & 0x00ff00ff) |
 			((((c0 & 0x0000ff00) + (c1 & 0x0000ff00)) >> 1) & 0x0000ff00) |